move setting animation callback to ScreenStackFragment's onCreateView
